Fix error when retrieving group members
It happens when there is a scale_group member object that
does not point to an actual guest.
The SL API returns an empty string for the referenced guest
object. The _to_node throws an exception for an empty string
input. Fix is to skip the _to_node call for empty members.