Enhance parse_error for OpenStack response.
Have parse_error to properly work for different json response formats.

parse_error properly parses nova API errors which returned in the
following similar format (array where its first element is a dictionary).

values = [{u'message': u'Invalid flavorRef provided.', u'code': 400}]

However, it fails to parse heat API errors which returned in the
following format.

values = [u'Conflict', u'There was a conflict when trying to complete your request.', 409,
        {u'message': u'The Stack (libcloud-group-1) already exists.',
         u'traceback': u'Traceback (most recent call last):\n\n
         File "/opt/stack/heat/heat/common/context.py", line 267, in wrapped\n
         return func(self, ctx, *args, **kwargs)\n\n  File "/opt/stack/heat/heat/engine/service.py",
         line 701, in create_stack\n    parent_resource_name)\n\n  File "/opt/stack/heat/heat/engine/service.py",
         line 598, in _parse_template_and_validate_stack\n    self._validate_new_stack(cnxt, stack_name, tmpl)\n\n
         File "/opt/stack/heat/heat/engine/service.py", line 563, in _validate_new_stack\n
         raise exception.StackExists(stack_name=stack_name)\n\nStackExists: The Stack (libcloud-group-1) already exists.\n',
         u'type': u'StackExists'}]

parse_error currently assumes that values[0] is a dictionary that contains
'code' and 'message' keys which is not always true. In some cases this dictionary
is positioned in other index then 0 and it might not have 'code' at all.

The enhancement would be to use self.status instead of the 'code' key
and to retrieve all the values for 'message' keys of the dictionaries (if exist).