Remove mandatory ex_region parameter from create_auto_scale_group.
create_auto_scale_group() of softlayer defines a mandatory parameter called ex_region.
ex_region denotes the region the auto-scale group will be created under and it is a
required parameter.

In libcloud, extended parameters (prefixed with 'ex') are defined as optional ones,
rather then mandatory, thus we should remove ex_region from softlayer's
create_auto_scale_group().

Looking into other libcloud drivers, it can be seen that they can be initialized with
a region (which is set to a default in case no value is being supplied).

We will follow a similar approach with SoftLayerAutoScaleDriver which will be initialized
with a region parameter (defaults to 'na-usa-east-1').

Auto-scale operations that require a region will be passed with the region the driver got
initialized with.